refactor: new directive API
